last refactor: fully functional bom api
